# spec/api/v1/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "meteroid",
    "description": "",
    "license": {
      "name": "AGPL-3.0-only",
      "identifier": "AGPL-3.0-only"
    },
    "version": "0.1.0"
  },
  "paths": {
    "/api/v1/addons": {
      "get": {
        "tags": [
          "Add-ons"
        ],
        "summary": "List add-ons",
        "operationId": "list_addons",
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "currency",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "include_archived",
            "in": "query",
            "description": "Include archived add-ons in the results (default: false)",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Page number (0-indexed)",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "per_page",
            "in": "query",
            "description": "Number of items per page",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "maximum": 100,
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of add-ons",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddOnListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "post": {
        "tags": [
          "Add-ons"
        ],
        "summary": "Create an add-on",
        "operationId": "create_addon",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAddOnRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Add-on created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddOn"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/api/v1/addons/{addon_id}": {
      "get": {
        "tags": [
          "Add-ons"
        ],
        "summary": "Get add-on details",
        "operationId": "get_addon",
        "parameters": [
          {
            "name": "addon_id",
            "in": "path",
            "description": "Add-on ID",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/AddOnId"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Add-on details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddOn"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Add-on not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "patch": {
        "tags": [
          "Add-ons"
        ],
        "summary": "Update an add-on",
        "operationId": "update_addon",
        "parameters": [
          {
            "name": "addon_id",
            "in": "path",
            "description": "Add-on ID",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/AddOnId"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAddOnRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Add-on updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddOn"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Add-on not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/api/v1/addons/{addon_id}/archive": {
      "post": {
        "tags": [
          "Add-ons"
        ],
        "summary": "Archive an add-on",
        "operationId": "archive_addon",
        "parameters": [
          {
            "name": "addon_id",
            "in": "path",
            "description": "Add-on ID",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/AddOnId"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Add-on archived"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Add-on not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/api/v1/addons/{addon_id}/unarchive": {
      "post": {
        "tags": [
          "Add-ons"
        ],
        "summary": "Unarchive an add-on",
        "operationId": "unarchive_addon",
        "parameters": [
          {
            "name": "addon_id",
            "in": "path",
            "description": "Add-on ID",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/AddOnId"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Add-on unarchived"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Add-on not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "AddOn": {
        "type": "object",
        "required": [
          "id",
          "name",
          "product_id",
          "price_id",
          "self_serviceable",
          "created_at"
        ],
        "properties": {
          "archived_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "fee_type": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ProductFeeTypeEnum"
              }
            ]
          },
          "id": {
            "$ref": "#/components/schemas/AddOnId"
          },
          "max_instances_per_subscription": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "price_id": {
            "$ref": "#/components/schemas/PriceId"
          },
          "product_id": {
            "$ref": "#/components/schemas/ProductId"
          },
          "self_serviceable": {
            "type": "boolean"
          }
        }
      },
...
      "AddOnId": {
        "type": "string",
        "format": "MeteroidId",
        "examples": [
          "add_7n42DGM5Tflk9n8mt7Fhc7"
        ]
      },
      "AddOnListResponse": {
        "type": "object",
        "required": [
          "data",
          "pagination_meta"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AddOn"
            }
          },
          "pagination_meta": {
            "$ref": "#/components/schemas/PaginationResponse"
          }
        }
      },
...
      "CreateAddOnRequest": {
        "type": "object",
        "required": [
          "name",
          "product_id",
          "price_id"
        ],
        "properties": {
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "max_instances_per_subscription": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "price_id": {
            "$ref": "#/components/schemas/PriceId"
          },
          "product_id": {
            "$ref": "#/components/schemas/ProductId"
          },
          "self_serviceable": {
            "type": "boolean"
          }
        }
      },
...
      "ErrorCode": {
        "type": "string",
        "enum": [
          "BAD_REQUEST",
          "NOT_FOUND",
          "CONFLICT",
          "FORBIDDEN",
          "UNAUTHORIZED",
          "INTERNAL_SERVER_ERROR"
        ]
      },
...
      "PaginationResponse": {
        "type": "object",
        "required": [
          "page",
          "per_page",
          "total_items",
          "total_pages"
        ],
        "properties": {
          "page": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "per_page": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "total_items": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "total_pages": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          }
        }
      },
...
      "PriceId": {
        "type": "string",
        "format": "MeteroidId",
        "examples": [
          "pri_7n42DGM5Tflk9n8mt7Fhc7"
        ]
      },
...
      "ProductFeeTypeEnum": {
        "type": "string",
        "enum": [
          "RATE",
          "SLOT",
          "CAPACITY",
          "USAGE",
          "EXTRA_RECURRING",
          "ONE_TIME"
        ]
      },
      "ProductId": {
        "type": "string",
        "format": "MeteroidId",
        "examples": [
          "prd_7n42DGM5Tflk9n8mt7Fhc7"
        ]
      },
...
      "RestErrorResponse": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "UpdateAddOnRequest": {
        "type": "object",
        "properties": {
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "max_instances_per_subscription": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "name": {
            "type": [
              "string",
              "null"
            ]
          },
          "price_id": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/PriceId"
              }
            ]
          },
          "self_serviceable": {
            "type": [
              "boolean",
              "null"
            ]
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}